Modify how winner is calculated - no longer hardcoded!
Winner is now calculated using new scoreState field in the history
objects. scoreState stores the 'scores' of each player's symbol in
each row, column and diagonals. p1 increases the 'score', p2 decreases
the 'score'.

Winner exists when there is abs('boardSize') number of one symbol in a
row/col/diagonal. In that case, calcuate the square numbers that
contributed to the win in term of the row/col/diagonal number, and the
board size.